New utility method ImgUtil.copy(RandomAccessibleInterval, IterableInterval, int) for multi-threaded copying with `int` threads.
Addresses the issue of copying "images" that are actually transformed
views using linear interpolators and others, which can be costly
and, from a scripting framework, it is not entirely trivial to
partition a RandomAccessibleInterval for multi-threaded processing.
While it could be done via Views.interval, it is simpler to
use a utility method to do it all.